Move `usePaymentActivityData()` to `PaymentActivity` because filters will need it.
